Refactor VisibilitySelectPresentation to be a dumb component
