+++ frontend/app/assets/javascripts/components/VisibilitySelect/VisibilitySelectData.js
@@ -1,40 +1,59 @@
 import PropTypes from 'prop-types'
 //FPCC
-import useVisibility from 'DataSource/useVisibility'
+import { useState, useEffect } from 'react'
 
 /**
- * @summary VisibilitySelectData
+ * @summary VisibilitySelectData - uses the state of a document to determine its visibility and handles the LOCAL state of the VisibilitySelect component.
  * @version 1.0.1
  * @component
  *
  * @param {object} props
  * @param {function} props.children
+ * @param {string} props.docId
+ * @param {string} props.docState
+ * @param {function} props.handleVisibilityChange
  *
  */
-function VisibilitySelectData({ children }) {
-  const { updateVisibilityToTeam, updateVisibilityToMembers, updateVisibilityToPublic } = useVisibility()
+function VisibilitySelectData({ children, docId, docState, handleVisibilityChange }) {
+  const [visibility, setVisibility] = useState('')
 
-  const handleVisibilityChange = (visibility, id) => {
-    switch (visibility) {
-      case 'team':
-        return updateVisibilityToTeam(id)
-      case 'members':
-        return updateVisibilityToMembers(id)
-      case 'public':
-        return updateVisibilityToPublic(id)
+  useEffect(() => {
+    setVisibility(convertStateToVisibility(docState))
+  }, [])
+
+  const handleChange = (event) => {
+    const newVisibility = event.target.value
+    setVisibility(newVisibility)
+    handleVisibilityChange(newVisibility, docId)
+  }
+
+  function convertStateToVisibility(state) {
+    switch (state) {
+      case 'New':
+        return 'team'
+      case 'Disabled':
+        return 'team'
+      case 'Enabled':
+        return 'members'
+      case 'Published':
+        return 'public'
       default:
-        return null
+        return ''
     }
   }
 
   return children({
-    handleVisibilityChange,
+    handleChange,
+    visibility,
   })
 }
 // PROPTYPES
-const { func } = PropTypes
+const { func, string } = PropTypes
 VisibilitySelectData.propTypes = {
   children: func,
+  docId: string,
+  docState: string,
+  handleVisibilityChange: func,
 }
 
 export default VisibilitySelectData
